refactor: rename and reorganize constants and components
